Backend-API/requirements.txt
Flask==3.0.0
pyArango==2.0.2
python-dotenv==1.0.0
flask-cors==4.0.0
flask-jwt-extended==4.5.3
bcrypt==4.1.1